Refactor completeTask function
